fix: rename package name to cfazure
